+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Define new TTP::Path::checkLength() function with per-OS configuration
     - http.pl compare takes advantage of new checkLength() function when creating dump files
     - HTTP::Compare::Browser no more considers linux-specific arguments
+    - Align site.example/Commands/backup-monitor-daemon.pl to new semantics
 
 ### 4.24.0